cc: Allow micro benchmarks to receive arguments from js.
This patch adds the ability for us to pass parameters via telemetry
or via javascript to the micro benchmarks. This allows us to be a bit
more flexible in what we test, as opposed to have a set of fixed
parameters to test.

BUG=307756
R=enne@chromium.org,nduca@chromium.org

Review URL: https://codereview.chromium.org/27739002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 752e2cf0862cd3daa0c219e8dfb13ebd5330e081