Fix chrome_tests.sh for local workflow
This CL adds testing for the case that the user has not specified -b and
--target values before operations related to them, since later on the
the script handles these values not being specified. This usage is
perscribed in developer docs at
http://www.chromium.org/developers/how-tos/using-valgrind. Additionally
I reworked the check in the assert to properly handle the case when the
dir passed into -b ends in a /, since I did that while trying to
understand the code and was very confused why the assert wasn't
tripping.

BUG=308576
TEST=Tested manually using a variety of different command lines to make
     sure the script behaves as I expected.

Review URL: https://codereview.chromium.org/26868006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 955429ac48d21b7184335eead3dc4506d9c3c2d8