Add a metric to measure the keyboard load time.
This involves adding a new extension api that's a callback called by the Keyboard code when all keyboard elements are loaded. This can be used for other purposes in the future, such as allowing the ash/aura side to show a loading animation, or handle showing of the keyboard differently.

BUG=265326
TEST=Start Chrome OS and focus on a text box and wait for the keyboard to appear. Then go to chrome://histograms and ensure that Keyboard.FirstLoadTime appears and the measurement is roughly how long it took for the keyboard to fully appear after focusing the text box.

Review URL: https://codereview.chromium.org/37213002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 73f05699a44e5756fc1479de100a8e60d1233006