This CL changes the UI for the malicious download notification in the download shelf. The change only affects downloads in the IsMalicious category.
Previously, the button in the notification said 'discard'. Now it says 'dismiss'.

(Note that this is overriden for DANGEROUS_HOST when the reporting feature is turned on. When the reporting feature is turned on, the button will still say "Discard & report".)

BUG=305387
TEST=
1. Download a malicious file, e.g. http://download.safebrowsingtest.com/download/test
2. Click the 'dismiss' button in the notification
3. The notification should disappear
4. Go to chrome://downloads and check that the download still appears there

Review URL: https://codereview.chromium.org/42363003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 19d4dd5a645fe4e593af61dddef743d0d88b61a3