This is the first step of a refactor I'm going to do to separate filling out the target and connecting the dependency tree. I'm hoping this refactor will dramatically simplify the threading code and fix some flakiness.
This change is mostly mechanical and replaces the lists of targets and config in the target object with a new pair that consists of a label and the pointer. Currently, they're still filled out at the same time.

The new structure also tracks the origin of the thing which simplifies some of the tracking in the describe function.

R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/48523006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68d1dd345e06e259a086c2ccc186f7f260bbf0d0