Abort prerenders which try to create two navigation entries.
We do not currently patch up the history correctly. Let's try to fix this by
just aborting prerenders which do this, rather than complicating the history
merge function. UMA will tell us if this is a problem and if this case turns
out to be important.

This does not affect the following:
- Calls to location.replace.
- <meta http-equiv=refresh> redirects with a timeout under 1 second.
- Code which sets location.href before the load event.

BUG=305660
TEST=PrerenderBrowserTest.PrerenderNewNavigationEntry
     PrerenderBrowserTest.PrerenderLocationReplaceNavigateToFirst
     PrerenderBrowserTest.PrerenderLocationReplaceNavigateToSecond

Review URL: https://codereview.chromium.org/68213005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ab5871bc50be5c7d0c42ed7b33052f0c5617ffaf