[Telemetry] Don't try to use symbols from another browser for stacks.
On the bots, the reference builds were trying to symbolize with the target
build's symbols. The problem was that we were searching for symbols in the
directory of the minidump_stackwalk binary instead of in the directory of the
browser.

BUG=280584

Review URL: https://codereview.chromium.org/75133008

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 95b0d33eb8b7fe6ca598dca50d5174d8800adae7