Get foreground tab on Android
This works around the fact that document.hidden doesn't work
reliably on Android, by assuming that the first tab in the tab
list is the foreground tab. This is always true at present, but is
not guaranteed in future.

BUG=

Review URL: https://codereview.chromium.org/83303003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 19b455af83ffc6f76142d86905da4f0ffeb3b544