Auto-updating script for Windows toolchain
Updates third_party/win_toolchain to pull down a toolchain
based on SHA1. The intention is that this will be a DEPS hook
once it's ready for more broad usage and the rest of
infrastructure is ready.

R=maruel@chromium.org
TBR=cpu@chromium.org
BUG=323300

Review URL: https://codereview.chromium.org/70493006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c46460594b7b0d5b6edde48df3bf08c6fc389bba