Telemtry: Remove '-1' when computing the scrollHeight for smooth scrolling.
We're currently only scrolling to "scrollHeight - 1" in Telemetry. Supposedly
this is to "allow for rounding errors on scaled viewports" but it means we'll
never scroll the entire page, which seems to cause problems for the
testScrollAction unit test.

BUG=323249

Review URL: https://codereview.chromium.org/99413002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3ff6324d1d4071b8932fa494f196fef64ff6eb3c