Infobar system refactor.
This changes the ownership model of infobars so that InfoBars are long-lived and
own their InfoBarDelegates directly.  The InfoBarService pseudo-owns the
InfoBars (instead of deleting them directly, it tells them when they're unowned
and expects them to delete themselves).

This fixes leaks when infobars are closed while not visible (e.g. in a
background tab) and in general makes the system clearer and easier to reason
about.

BUG=62154
TEST=none
R=erg@chromium.org, sail@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/22694006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 42b0ecedf7eb4beea125ac17ffc0cd373f4b420c