[Telemetry] Fix page_cycler for pages with client side redirects.
Page.disable clears all scripts to evaluate on commit. There is no way
to detect client side redirects because linkedin, for example, performs
its client side redirect after the page has fully loaded.

In order for the page_cyclers to work on pages with client side
redirects, the script to evaluate on commit must execute on the
redirected page.

So to workaround this, we avoid ever calling Page.disable. I don't
expect continuing to listen for notifications here to impact
performance, but if we do see any performance reduction on the bots we
should revert and search for another approach.

This was originally landed as r209505 but had to be rolled out. This
patch simply updates the original patch since the regression was
unrelated and has been fixed.

BUG=253604
NOTRY=true
TEST=tools/perf/run_measurement --browser=canary page_cycler tools/perf/page_sets/top_10_mobile.json --pageset-repeat=2

Review URL: https://codereview.chromium.org/103373003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 31e81c4edae80e91b9904e79cce4364f33ac7272