[Android] Fix session_restore tests on occam svelte
The session restore tests are failing on occam svelte because it
doesn't have enough memory to load one of the background tabs. Fix
this by catching the timeout when trying to access the tab, and
ignoring the results for that tab.

BUG=

Review URL: https://codereview.chromium.org/103963004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 46a20dedad3c121fa8368641d01962f20b21f79f