[Telemetry] Add thread duration for Async slices
An async event can theoritically start on a thread and end on another
but in practice that does not happen.

This CL calculates the thread_duration for async slices that started
and ended on the same thread.

BUG=329059

Review URL: https://codereview.chromium.org/112593004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a4c18621db8e381e20627627cabf962d730419f9