[Telemetry]: Move assert on page measurement results name into Value object.
The use of a '.' in a trace name only breaks the telemetry value result if it
occurs in a value which does not have a trace_name.  This CL moves the assert
into the Value object so that the assert doesn't fire in other situations
(such as the media.tough_media_cases benchmark), where the use of the '.' is
actually ok.

BUG=332041,329845

Review URL: https://codereview.chromium.org/126013002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0d9d94181a62e94980ebd41ae19fbc5250a12f02