Add extension install verification entries to histograms.xml
These histograms got renamed and reworked in
https://codereview.chromium.org/132373002 to be more useful, but the existing
ones are already in the M33 branch and we'd like to be able to see their
results. Later on we'll be able to remove these ones, and that work is tracked
by the bug referenced below.

BUG=333934
R=asvitkine@chromium.org

Review URL: https://codereview.chromium.org/136453006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e23927d3d78afed3a3d45cc041c41d0fb633c409