Telemetry: Fix timeline frame counting on Mac
We want to count the last the last swap in the embedding
chain. Usually there is a hardware swap for this but on mac
there doesn't appear to be one. This uses the decoder swap
for now until we know the canonical swap call and add a trace
there.

TBR=nduca@chromium.org
NOTRY=true

No-try since this is blocking this from working on the mac and I've run the tests locally.

Review URL: https://codereview.chromium.org/141163004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6796617515596dad1d4b71cc5903000175ab7feb