Re-Reland: [Telemetry] Properly handle retries after browser crash.
Previously it was easy for the results to get into a bad state and assert in a
variety of ways.

This ensures we ignore the partial results in the case of a crash.

This reland is tested with GPU functional tests. The main fix was to make all
result types shallow copyable.

BUG=
TBR=tonyg@chromium.org

Review URL: https://codereview.chromium.org/147853002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 036ed29329b70aff503a2e8731ee452ea27ddba0