Add policy that forces SAML users to log in online periodically
This CL adds a user policy which limits the time for which a user
authenticated via SAML can log in offline. When the limit is reached, the
user is required to go through the online flow on next login.

BUG=332081
TEST=Full unit test and browser test coverage

Review URL: https://codereview.chromium.org/143463009

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3b9deb40bbf49a1b5f186293f28dc519b51a8995