Verify variations seed signature and log results in a histogram.
Modifies VariationsSeedStore to log histograms with the result of
verifying the variations seed against the signature returned from
the server. One histogram is logged when verifying the seed
on startup and another when it's received and stored.

BUG=336536

Review URL: https://codereview.chromium.org/148153015

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2a6156dc5ebee4b091022603cb9250a67ef005d9