Remove line that crashes image_decoding.tough_decoding_cases
image_decoding.py was calling PageMeasurementResults.Add with a value
of the wrong type (a string instead of a scalar type). This was
crashing when the type was checked.

Note: I am not sure why this path is being followed in 
this test, but the code on this path was clearly broken.

NOTRY=true
TBR=tonyg@chromium.org
BUG=338813

Review URL: https://codereview.chromium.org/146283005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2b85e7e2baf47a90149e009a7eff35da1eca4246