[GCM] Enable Registration Request to give after 5 retries.
When a Chrome app issues a registration, that operation might be looping endlessly, when an error is recoverable. If meanwhile app wants to issue another registration (perhaps with a different set of senders) it will always be receiving and error (asynchronous operation pending) without a chance to cancel.

The update is going to allow the registraiton to fail, after a 6 attempts (1 initial attempt + 5 retries). That approach ensures that applications follow an exponential backoff policy and have a chance to issue a new registration request after a period of time.

BUG=344010

Review URL: https://codereview.chromium.org/180113005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 00185c40e4d97dac2081389bf3784c350c92c6c2