[telemetry] Turn on GPU benchmarking extension for all runs
This patch turns on GPU benchmarking extension for all runs [unless opted out via BrowserOptions]. This is done mainly to simplify the actions system. Right now, we have to programmatically inspect a pageset's actions in order to determine if it might
eventually scroll or do a synthetic gesture, and then if it will, we add in the gpu
benchmarking extension. This is not possible when pagesets are python-based.

In reality, the only side effect of the extension being on is that render stats are tracked. We believe this to be effectively free: rather than trying to add more complexity into the actions and pageset system in order to figure out if gestures are needed, I propose we try moving this to always be on and see if we can get by. The intent in the long term is to make gesture creation available on devtools or in the platform, anyway, so this seems like a low architectural risk.

Review URL: https://codereview.chromium.org/189463007

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dfe9297190a2bbb319274a59303d32af2ea1be94