Add histogram for consumer use of account whitelisting
We know that enterprise managed Chrome OS devices often
allow accounts from only their domain to be used on the
device. We have functionality to allow the Owner of a
consumer device to do similar account whitelisting.
We'd like to track consumer usage of this feature, so add
a histogram for it.

BUG=chromium:345352
TEST=build the browser

Review URL: https://codereview.chromium.org/193143002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ca12435f2c5de9de9d4c54bd91f9caad571c97e5