[telemetry] run_benchmark for measurements
If the argument is a PageTest and a PageSet, wrap it in a Test and continue running normally.
Most of the work is modifying the test_runner to print both Tests and PageTests in the test list.

Auxiliary changes:
- Make more measurements private, so they won't show up in the run_benchmark list.
- Remove PageTest.CreatePageSet(), it's superseded by Test.CreatePageSet().


BUG=352799
TEST=Run it!

Review URL: https://codereview.chromium.org/202573002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ad73ae7dfcc59eaa2d8080455fba69d5b384842e