Wire up the mmap profiler into the memory_inspector frontend.
This makes possible to get an aggregated view of a mmap dump.
The change introduces also three new /profile/ endpoints to the server
which allow to create a profile (for a recent mmap dump) and later
grab the data for the charts corresponding to that profile.

BUG=340294
NOTRY=true

Review URL: https://codereview.chromium.org/195893027

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 22a449259d7582c6f9fcaa338bf46e807c146059