Adds suppression for leak
The objects showing up as being leaked (CallbackHolder) are embedded
in the v8 heap. This tests seems to be shutting down in a way that
doesn't cause the finalizers for these objects to run and we
leak. Temporarily supressing until we sort out the issue.

BUG=356785
TEST=none
R=jochen@chromium.org
TBR=jochen@chromium.org

Review URL: https://codereview.chromium.org/213723003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0b6e719e5e89b384c326f97919647ac96f51f5e0