telemetry: handle tests that are not supported on a platform.
When a test or measurement was not supported on a platform, we used to print
a warning and call sys.exit(0). This caused problems with unittests, because
it was considered a failure (for page measurements it was not a failure).

This patch introduces the page_test.TestNotSupportedOnPlatformFailure
exception. Tests and measurements throw this exception if they can't run on
a platform. PageTestRunner and unit tests catch these exceptions and
print a warning, but they don't consider it an error.

R=tonyg@chromium.org, dtu@chromium.org, vmpstr@chromium.org
BUG=

Review URL: https://codereview.chromium.org/100443007

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e0e89d95cb6c70929036bbd59f124914a3a2f5ab