Add Brett as the OWNER of checkdeps code.
He was the author of this tool and with Joi leaving Chromium he seems to
be a good candidate. Others can be added later if needed.

This is good for people changing this tool, so they know now who to ask for
reviews.

TBR=brettw@chromium.org
NOTRY=true

Review URL: https://codereview.chromium.org/227013002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5d41ef337319bdff32d745f11d7c430deb74778f