Add a new dotfile command line argument to GN
This allows one to specify the .gn file manually when also overriding the
--root source path.

This removes the --secondary command line argument. There's no use case to
specify this on the command line rather than in the .gn file.

Improves some error reporting if the root path isn't found.

Fixes a crash in scope tear-down. When crating a closure or doing a scope merge, the template pointers would be copied rather than the template objects, but the scope expects to own the pointers. This would result in a double-delete. The solution is to create a new copy of the template in the Scope::NonRecursiveMergeTo

BUG=
R=noelallen@chromium.org

Review URL: https://codereview.chromium.org/218403014

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0485cf05e88915e1b175a02d60b54e5c1af31a44