Telemetry: Skip page if a page action is not supported by the browser.
If a page requires touch scrolling, we currently report a failure if the browser
doesn't support touch. This means that no average numbers will be reported on the
dashboard.
Instead of failing, we could skip the page and show a warning message to make the
user aware of that.

BUG=355661

Review URL: https://codereview.chromium.org/230373005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 740d029a927e258b35a547bc41fa5f4fd4d59141