Make isolate_driver.py process build.ninja and extract dependencies.
This uses a few assumption:
- This basically breaks non-ninja build for component builds. This never worked
  anyway.
- This assumes the file format of .ninja files. This will likely be quite
  obvious when this breaks.
- It makes some assumptions about the build steps, for example '.so.TOC' ->
  '.so'.

On the other hand, it creates a deterministic dependency tree, which is awesome.
Technically it would work as well for non-component builds but I don't want to
go this far yet. But in the end, that's the goal that nobody has to enter the
binary dependencies in the .isolate files.

R=csharp@chromium.org
BUG=360223,333473

Review URL: https://codereview.chromium.org/228463003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f4b59e75dd0924d064010d18687d62bb51b82642