[Telemetry] Allow android_commands to be imported on all platforms.
The psutil dependency was the only thing preventing this from working
everywhere. We didn't even need the functionality that it was providing.

This brings us a step closer to working with a mac host and android device.

BUG=355646

Review URL: https://codereview.chromium.org/242693002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 989a32e34cb3e74950e3e98c32a973eb2c8181cb